# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RESTful API fullstack challenge",
        "description": "This is the API documentation for the application",
        "contact": {
            "email": "higorribeiroaraujo007@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/phone": {
            "get": {
                "tags": [
                    "Phones"
                ],
                "summary": "Get all phones",
                "description": "Get all phones",
                "operationId": "c84750a44448a2d742dec9cfbdff7d04",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "number of page",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Phones"
                ],
                "summary": "Create a new phone",
                "description": "Create a new phone by providing the necessary information",
                "operationId": "68f792baecce3adab0a1555462371c28",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "num"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "The user id",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "num": {
                                        "description": "The number of phone",
                                        "type": "string",
                                        "example": "(99) 99999-9999"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Phone created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "num": {
                                            "type": "string",
                                            "example": "(99) 99999-9999"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/phone/{id}": {
            "get": {
                "tags": [
                    "Phones"
                ],
                "summary": "Get phone",
                "description": "Get phone",
                "operationId": "4c6848d41f8137b980909c368de44ce0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "phone id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Phones"
                ],
                "summary": "Delete phone",
                "description": "Delete phone",
                "operationId": "53012e45008fbff7f68151b6c61d85b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "phone id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Phone removed successfully.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Phones"
                ],
                "summary": "Update a new phone",
                "description": "Update a new phone by providing the necessary information",
                "operationId": "044f33be4999c05dde0227bc993c8a5e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "phone id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "The user id",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "num": {
                                        "description": "The number of phone",
                                        "type": "string",
                                        "example": "(99) 99999-9999"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Phone updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "num": {
                                            "type": "string",
                                            "example": "(99) 99999-9999"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}